+++ examples/Book/JsonApi/Document/BookDocument.php
@@ -58,8 +58,7 @@ class BookDocument extends AbstractSingleResourceDocument
      */
     public function getLinks()
     {
-        return new Links(
-            "http://example.com/api",
+        return Links::createWithoutBaseUri(
             [
                 "self" => new Link("/books/" . $this->getResourceId())
             ]