+++ CHANGELOG.md
@@ -9,8 +9,12 @@ CHANGED:
 
 - Renamed `getDefaultRelationships()` to `getDefaultIncludedRelationships()` in transformers to better reflect its meaning
 - The "data" key of relationships won't be present in the response when it is empty
+- Renamed `Links::addLinks()` to `Links::setLinks` and `Links::addLink()` to `Links::setLink()`
 
 REMOVED:
+
+- Most of the `Links::create*` static methods to simplify creation
+- `RelativeLinks` class as it became useless
 
 FIXED: